api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
alembic
redis
pydantic-settings
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-multipart
